+++ Source/Podio .NET/Services/TaskService.cs
@@ -73,7 +73,7 @@ namespace PodioAPI.Services
         /// <param name="refId"></param>
         /// <param name="limit"></param>
         /// <returns></returns>
-        public async Task<TaskSummary> GetTaskSummaryForReference(string refType, int refId, int limit = 4)
+        public async Task<TaskSummary> GetTaskSummaryForReference(string refType, long refId, int limit = 4)
         {
             string url = string.Format("/task/{0}/{1}/summary", refType, refId);
             var requestData = new Dictionary<string, string>()
@@ -107,7 +107,7 @@ namespace PodioAPI.Services
         /// <param name="refType"></param>
         /// <param name="refId"></param>
         /// <returns></returns>
-        public async Task<int> GetTaskCount(string refType, int refId)
+        public async Task<int> GetTaskCount(string refType, long refId)
         {
             string url = string.Format("/task/{0}/{1}/count", refType, refId);
             dynamic response = await _podio.Get<dynamic>(url);
@@ -131,7 +131,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="taskId"></param>
         /// <returns></returns>
-        public async Task<Models.Task> GetTask(int taskId)
+        public async Task<Models.Task> GetTask(long taskId)
         {
             string url = string.Format("/task/{0}", taskId);
             return await _podio.Get<Models.Task>(url);
@@ -165,7 +165,7 @@ namespace PodioAPI.Services
         ///     If set to true, the object will not be bumped up in the stream and notifications will not be
         ///     generated. Default value: false
         /// </param>
-        public async System.Threading.Tasks.Task IncompleteTask(int taskId, bool hook = true, bool silent = false)
+        public async System.Threading.Tasks.Task IncompleteTask(long taskId, bool hook = true, bool silent = false)
         {
             string url = string.Format("/task/{0}/incomplete", taskId);
             url = Utility.PrepareUrlWithOptions(url, new CreateUpdateOptions(silent, hook));
@@ -185,7 +185,7 @@ namespace PodioAPI.Services
         ///     generated. Default value: false
         /// </param>
         /// <returns></returns>
-        public async Task<List<Models.Task>> CreateTask(TaskCreateUpdateRequest task, string refType = null, int? refId = null,
+        public async Task<List<Models.Task>> CreateTask(TaskCreateUpdateRequest task, string refType = null, long? refId = null,
             bool hook = true, bool silent = false)
         {
             string url = "/task/";
@@ -227,7 +227,7 @@ namespace PodioAPI.Services
         /// </param>
         /// <returns></returns>
         public async Task<List<Models.Task>> CreateTask(string text, DateTime? dueDate = null, string description = null,
-            int? responsible = null, bool isPrivate = true, string refType = null, int? refId = null, bool hook = true,
+            long? responsible = null, bool isPrivate = true, string refType = null, long? refId = null, bool hook = true,
             bool silent = false)
         {
             var task = new TaskCreateUpdateRequest
@@ -282,7 +282,7 @@ namespace PodioAPI.Services
         ///     <para>API Reference: https://developers.podio.com/doc/tasks/delete-label-151302 </para>
         /// </summary>
         /// <param name="labelId"></param>
-        public async System.Threading.Tasks.Task DeleteLabel(int labelId)
+        public async System.Threading.Tasks.Task DeleteLabel(long labelId)
         {
             string url = string.Format("/task/label/{0}", labelId);
             await _podio.Delete<dynamic>(url);
@@ -295,7 +295,7 @@ namespace PodioAPI.Services
         /// <param name="taskId"></param>
         /// <param name="beforeTaskId"></param>
         /// <param name="afterTaskId"></param>
-        public async System.Threading.Tasks.Task RankTask(int taskId, int beforeTaskId, int afterTaskId)
+        public async System.Threading.Tasks.Task RankTask(long taskId, int beforeTaskId, int afterTaskId)
         {
             string url = string.Format("/task/{0}/rank", taskId);
             dynamic requestData = new
@@ -311,7 +311,7 @@ namespace PodioAPI.Services
         ///     <para>API Reference: https://developers.podio.com/doc/tasks/remove-task-reference-6146114 </para>
         /// </summary>
         /// <param name="taskId"></param>
-        public async System.Threading.Tasks.Task RemoveTaskReference(int taskId)
+        public async System.Threading.Tasks.Task RemoveTaskReference(long taskId)
         {
             string url = string.Format("/task/{0}/ref", taskId);
             await _podio.Delete<dynamic>(url);
@@ -324,7 +324,7 @@ namespace PodioAPI.Services
         /// <param name="labelId"></param>
         /// <param name="text">The name of the new label</param>
         /// <param name="color">The color of the label in hex format (xxxxxx)</param>
-        public async System.Threading.Tasks.Task UpdateLabel(int labelId, string text, string color)
+        public async System.Threading.Tasks.Task UpdateLabel(long labelId, string text, string color)
         {
             string url = string.Format("/task/label/{0}", labelId);
             dynamic requestData = new
@@ -343,7 +343,7 @@ namespace PodioAPI.Services
         /// <param name="description"></param>
         /// <param name="hook"></param>
         /// <param name="silent"></param>
-        public async System.Threading.Tasks.Task UpdateTaskDescription(int taskId, string description, bool hook = true, bool silent = false)
+        public async System.Threading.Tasks.Task UpdateTaskDescription(long taskId, string description, bool hook = true, bool silent = false)
         {
             string url = string.Format("/task/{0}/description", taskId);
             url = Utility.PrepareUrlWithOptions(url, new CreateUpdateOptions(silent, hook));
@@ -381,7 +381,7 @@ namespace PodioAPI.Services
         /// <param name="text">The new text of the task</param>
         /// <param name="hook"></param>
         /// <param name="silent"></param>
-        public async System.Threading.Tasks.Task UpdateTaskText(int taskId, string text, bool hook = true, bool silent = false)
+        public async System.Threading.Tasks.Task UpdateTaskText(long taskId, string text, bool hook = true, bool silent = false)
         {
             string url = string.Format("/task/{0}/text", taskId);
             url = Utility.PrepareUrlWithOptions(url, new CreateUpdateOptions(silent, hook));
@@ -401,7 +401,7 @@ namespace PodioAPI.Services
         /// <param name="dueDateTime">The date and time the task is due (in local date and time)</param>
         /// <param name="hook"></param>
         /// <param name="silent"></param>
-        public async System.Threading.Tasks.Task UpdateTaskDueOn(int taskId, DateTime dueOn, DateTime dueDateTime, bool hook = true,
+        public async System.Threading.Tasks.Task UpdateTaskDueOn(long taskId, DateTime dueOn, DateTime dueDateTime, bool hook = true,
             bool silent = false)
         {
             string url = string.Format("/task/{0}/due", taskId);
@@ -421,7 +421,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="taskId"></param>
         /// <param name="labelIds"></param>
-        public async System.Threading.Tasks.Task UpdateTaskLabels(int taskId, List<int> labelIds)
+        public async System.Threading.Tasks.Task UpdateTaskLabels(long taskId, List<int> labelIds)
         {
             string url = string.Format("/task/{0}/label/", taskId);
             dynamic requestData = labelIds;
@@ -438,7 +438,7 @@ namespace PodioAPI.Services
         /// <param name="refId"></param>
         /// <param name="hook"></param>
         /// <param name="silent"></param>
-        public async System.Threading.Tasks.Task UpdateTaskReference(int taskId, string refType, int refId, bool hook = true, bool silent = false)
+        public async System.Threading.Tasks.Task UpdateTaskReference(long taskId, string refType, int refId, bool hook = true, bool silent = false)
         {
             string url = string.Format("/task/{0}/ref", taskId);
             url = Utility.PrepareUrlWithOptions(url, new CreateUpdateOptions(silent, hook));
@@ -578,7 +578,7 @@ namespace PodioAPI.Services
         ///     generated. Default value: false.
         /// </param>
         /// <returns></returns>
-        public async Task<int?> CompleteTask(int taskId, bool hook = true, bool silent = false)
+        public async Task<long?> CompleteTask(long taskId, bool hook = true, bool silent = false)
         {
             string url = string.Format("/task/{0}/complete", taskId);
             url = Utility.PrepareUrlWithOptions(url, new CreateUpdateOptions(silent, hook));
@@ -598,7 +598,7 @@ namespace PodioAPI.Services
         ///     generated. Default value: false.
         /// </param>
         /// <param name="responsible">The contact responsible (user_id), or null if no one should be responsible.</param>
-        public async System.Threading.Tasks.Task AssignTask(int taskId, int? responsible = null, bool silent = false)
+        public async System.Threading.Tasks.Task AssignTask(long taskId, long? responsible = null, bool silent = false)
         {
             string url = string.Format("/task/{0}/assign", taskId);
             url = Utility.PrepareUrlWithOptions(url, new CreateUpdateOptions(silent));
